# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "financialPortfolioApiBackend",
        "description": "API Documentation for your application.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Development Server"
        },
        {
            "url": "http://localhost",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/google/callback": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Handle Google authentication callback",
                "operationId": "586f999d90783dedb953928d5eba4d99",
                "responses": {
                    "200": {
                        "description": "Successful Google login, returns authentication token and user data.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Login com Google realizado com sucesso!"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "YOUR_SANCTUM_TOKEN"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/UserGoogleLoginResponse"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication failed with Google.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Falha na autenticação com Google."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Não foi possível autenticar com sua conta Google. Por favor, tente novamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error during user creation.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ocorreu um erro ao tentar criar o usuário via Google."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Por favor, tente novamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/transactions/reverse": {
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Reverse a completed transaction",
                "operationId": "be8340b36739c9d42fc039a11394f499",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReversalRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transaction reversed successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Reversal successful!"
                                        },
                                        "reversal_transaction": {
                                            "$ref": "#/components/schemas/TransactionResponse"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request, e.g., invalid transaction ID, already reversed, or business rule violation.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Falha ao realizar a reversão."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Transaction not found or not eligible for reversal."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - missing or invalid token.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthenticationError"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - user does not have permission to reverse this transaction.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedReversalError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error - invalid input data.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ocorreu um erro interno."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Detalhes técnicos do erro."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UserTypeEnum": {
                "title": "UserTypeEnum",
                "description": "Tipos de usuário permitidos no sistema",
                "type": "string",
                "enum": [
                    "COMMON",
                    "MERCHANT",
                    "ADMIN"
                ],
                "example": "COMMON"
            },
...
            "UserGoogleLoginResponse": {
                "title": "User Google Login Response",
                "description": "User data returned after successful Google login.",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "a1b2c3d4-e5f6-7890-1234-567890abcdef"
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john.doe@example.com"
                    },
                    "document": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "balance": {
                        "type": "number",
                        "format": "float",
                        "example": 0
                    },
                    "user_type": {
                        "$ref": "#/components/schemas/UserTypeEnum"
                    },
                    "google_id": {
                        "type": "string",
                        "example": "102345678901234567890",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-27T10:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-27T10:00:00.000000Z"
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "ReversalRequest": {
                "title": "Reversal Request",
                "description": "Data required to request a transaction reversal.",
                "required": [
                    "original_transaction_id",
                    "reason"
                ],
                "properties": {
                    "original_transaction_id": {
                        "description": "The UUID of the original transaction to be reversed. This transaction must exist.",
                        "type": "string",
                        "format": "uuid",
                        "example": "a1b2c3d4-e5f6-7890-1234-567890abcdef"
                    },
                    "reason": {
                        "description": "The reason for the reversal. Must be at least 10 characters long.",
                        "type": "string",
                        "maxLength": 255,
                        "minLength": 10,
                        "example": "Duplicate transaction due to system error."
                    }
                },
                "type": "object"
            },
...
            "AuthenticationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Falha na autenticação."
                    },
                    "error": {
                        "type": "string",
                        "example": "Credenciais inválidas."
                    }
                },
                "type": "object"
            },
            "TransactionResponse": {
                "description": "Schema representing a transaction, including details about sender and recipient.",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "d3f4e5c6-7890-1234-5678-abcdefabcdef"
                    },
                    "payer_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "a1b2c3d4-5678-90ef-1234-567890abcdef"
                    },
                    "payee_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "b2c3d4e5-6789-01ab-2345-678901abcdef"
                    },
                    "amount": {
                        "type": "string",
                        "format": "decimal",
                        "example": "150.00"
                    },
                    "type": {
                        "type": "string",
                        "example": "TRANSFER"
                    },
                    "status": {
                        "type": "string",
                        "example": "REVERSED"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-01T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-02T15:00:00Z"
                    }
                },
                "type": "object"
            },
            "UnauthorizedReversalError": {
                "description": "Returned when a user attempts to reverse a transaction they are not authorized to reverse.",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Você não tem permissão para reverter esta transação."
                    },
                    "error": {
                        "type": "string",
                        "example": "Reversão não autorizada."
                    }
                },
                "type": "object"
            },
...
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Erro de validação."
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}